+++ cmd/internal/cli/build.go
@@ -47,6 +47,7 @@ var buildArgs struct {
 	sandbox       bool
 	update        bool
 	nvidia        bool
+	nvccli        bool
 	rocm          bool
 	writableTmpfs bool // For test section only
 }
@@ -217,6 +218,16 @@ var buildNvFlag = cmdline.Flag{
 	Usage:        "inject host Nvidia libraries during build for post and test sections (not supported with remote build)",
 }
 
+// --nvccli
+var buildNvCCLIFlag = cmdline.Flag{
+	ID:           "buildNvCCLIFlag",
+	Value:        &buildArgs.nvccli,
+	DefaultValue: false,
+	Name:         "nvccli",
+	Usage:        "use nvidia-container-cli for GPU setup (experimental)",
+	EnvKeys:      []string{"NVCCLI"},
+}
+
 // --rocm
 var buildRocmFlag = cmdline.Flag{
 	ID:           "rocmFlag",
@@ -281,6 +292,7 @@ func init() {
 		cmdManager.RegisterFlagForCmd(&commonPEMFlag, buildCmd)
 
 		cmdManager.RegisterFlagForCmd(&buildNvFlag, buildCmd)
+		cmdManager.RegisterFlagForCmd(&buildNvCCLIFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildRocmFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildBindFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildWritableTmpfsFlag, buildCmd)