+++ pkg/util/singularityconf/config.go
@@ -42,6 +42,7 @@ type File struct {
 	AllowContainerDir       bool     `default:"yes" authorized:"yes,no" directive:"allow container dir"`
 	AllowContainerEncrypted bool     `default:"yes" authorized:"yes,no" directive:"allow container encrypted"`
 	AlwaysUseNv             bool     `default:"no" authorized:"yes,no" directive:"always use nv"`
+	UseNvCCLI               bool     `default:"no" authorized:"yes,no" directive:"use nvidia-container-cli"`
 	AlwaysUseRocm           bool     `default:"no" authorized:"yes,no" directive:"always use rocm"`
 	SharedLoopDevices       bool     `default:"no" authorized:"yes,no" directive:"shared loop devices"`
 	MaxLoopDevices          uint     `default:"256" directive:"max loop devices"`
@@ -309,6 +310,14 @@ allow container encrypted = {{ if eq .AllowContainerEncrypted true }}yes{{ else 
 # should be executed implicitly with the --nv option (useful for GPU only 
 # environments). 
 always use nv = {{ if eq .AlwaysUseNv true }}yes{{ else }}no{{ end }}
+
+# USE NVIDIA-NVIDIA-CONTAINER-CLI ${TYPE}: [BOOL]
+# DEFAULT: no
+# EXPERIMENTAL
+# If set to yes, Singularity will attempt to use nvidia-container-cli to setup
+# GPUs within a container when the --nv flag is enabled.
+# If no (default), the legacy binding of entries in nvbliblist.conf will be performed.
+use nvidia-container-cli = {{ if eq .UseNvCCLI true }}yes{{ else }}no{{ end }}
 
 # ALWAYS USE ROCM ${TYPE}: [BOOL]
 # DEFAULT: no